Create Basic Components for Quick Create Section
The components created here are placeholders for more robust design and
state.  Currently mocking up the components I will need for the MVP and
then will add state and interactivity as well as design.

The Quick Create section will be the top level way to create a task, but
users will also be able to click "more" if they want to add a set date or
frequency.  The defaults will be the current date and non-repeating.