Add State To App.js To Handle Checkbox Change
This commit introduces the handleClick method which controls any changes to the checkbox state.  Additionally, the state for CreatedTask was moved out of App and into CompletedTaskList.  This seemed more logical to me than having the state management so far away from the component that will use it. Not sure if this is best practices or not though.  We shall see during code review if a refactor is in order.